+++ src/GameClient.ts
@@ -1,6 +1,13 @@
-import { RunnerAdvanceConditionFunc, RunnerRenderingMode, RunnerV1, RunnerV2, RunnerV3 } from "@akashic/headless-driver";
-import { Canvas } from "canvas";
-import { EngineVersions } from "./types";
+import { RunnerV1, RunnerV2, RunnerV3 } from "@akashic/headless-driver";
+import { ResourceFactory as ResourceFactoryV1 } from "@akashic/headless-driver-runner-v1/lib/platform/ResourceFactory";
+import { ResourceFactory as ResourceFactoryV2 } from "@akashic/headless-driver-runner-v2/lib/platform/ResourceFactory";
+import {
+	NodeCanvasResourceFactory as ResourceFactoryV3_NodeCanvas
+} from "@akashic/headless-driver-runner-v3/lib/platform/NodeCanvasResourceFactory";
+import { NullResourceFactory as ResourceFactoryV3_Null } from "@akashic/headless-driver-runner-v3/lib/platform/NullResourceFactory";
+import type { EngineVersions } from "./types";
+import type { RunnerAdvanceConditionFunc, RunnerRenderingMode } from "@akashic/headless-driver";
+import type { Canvas } from "canvas";
 
 type Runner = RunnerV1 | RunnerV2 | RunnerV3;
 
@@ -11,6 +18,22 @@ export interface GameClientParameterObject<EngineVersion extends keyof EngineVer
 	game: EngineVersions[EngineVersion]["game"];
 	type: GameClientInstanceType;
 	renderingMode: RunnerRenderingMode;
+}
+
+export interface GameClientCreateImageAssetParameterObject {
+	id: string;
+	path: string;
+	width: number;
+	height: number;
+}
+
+export interface GameClientCreateAudioAssetParameterObject {
+	id: string;
+	path: string;
+	duration: number;
+	system: any;
+	loop: boolean;
+	hint: any;
 }
 
 /**
@@ -120,6 +143,42 @@ export class GameClient<EngineVersion extends keyof EngineVersions = keyof Engin
 		}
 
 		throw new Error(`GameClient#getPrimarySurface(): renderingMode "${mode}" is not supported`);
+	}
+
+	/**
+	 * ダミーの ImageAsset を生成する。
+	 * @param param ImageAsset の生成に必要なパラメータ。
+	 */
+	createDummyImageAsset(param: GameClientCreateImageAssetParameterObject): any {
+		const resFac = this.runner.platform.getResourceFactory();
+
+		if (resFac instanceof ResourceFactoryV1) {
+			return resFac.createImageAsset(param.id, param.path, param.width, param.height);
+		} else if (resFac instanceof ResourceFactoryV2) {
+			return resFac.createImageAsset(param.id, param.path, param.width, param.height);
+		} else if (resFac instanceof ResourceFactoryV3_NodeCanvas || resFac instanceof ResourceFactoryV3_Null) {
+			return resFac.createImageAsset(param.id, param.path, param.width, param.height);
+		}
+
+		throw Error("GameClient#createAudioAsset(): Could not create a image asset");
+	}
+
+	/**
+	 * ダミーの AudioAsset を生成する。
+	 * @param param AudioAsset の生成に必要なパラメータ。
+	 */
+	createDummyAudioAsset(param: GameClientCreateAudioAssetParameterObject): any {
+		const resFac = this.runner.platform.getResourceFactory();
+
+		if (resFac instanceof ResourceFactoryV1) {
+			return resFac.createAudioAsset(param.id, param.path, param.duration, param.system, param.loop, param.hint);
+		} else if (resFac instanceof ResourceFactoryV2) {
+			return resFac.createAudioAsset(param.id, param.path, param.duration, param.system, param.loop, param.hint);
+		} else if (resFac instanceof ResourceFactoryV3_NodeCanvas || resFac instanceof ResourceFactoryV3_Null) {
+			return resFac.createAudioAsset(param.id, param.path, param.duration, param.system, param.loop, param.hint);
+		}
+
+		throw Error("GameClient#createAudioAsset(): Could not create a audio asset");
 	}
 
 	/**