+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## unreleased changes
+* `GameContext#createDummyImageAsset()`, `GameContext#createDummyAudioAsset()` を追加
+
 ## 2.0.0
 * headless-driver@1.7.0 に追従
 * `GameClientStartParameterObject#renderingMode` を追加